api-gateway/pyproject.toml
[tool.poetry]
name = "intelligence-service-api-gateway"
version = "1.0.0"
description = "API Gateway pour le système Intelligence-Service"
authors = ["Intelligence Service Team <admin@intelligence-service.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
httpx = "^0.25.2"
python-dotenv = "^1.0.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
prometheus-client = "^0.19.0"
redis = "^5.0.1"
opentelemetry-api = "^1.21.0"
opentelemetry-sdk = "^1.21.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-exporter-otlp = "^1.21.0"
orjson = "^3.9.10"
jwt = "^1.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"